+++ nexus-repository-conda-it/src/test/java/org/sonatype/nexus/plugins/conda/internal/fixtures/CondaRepoRecipes.groovy
@@ -1,0 +1,27 @@
+package org.sonatype.nexus.plugins.conda.internal.fixtures
+
+import javax.annotation.Nonnull
+
+import org.sonatype.nexus.repository.Repository
+import org.sonatype.nexus.repository.config.Configuration
+import org.sonatype.nexus.testsuite.testsupport.fixtures.ConfigurationRecipes
+
+import groovy.transform.CompileStatic
+
+
+/**
+ * Factory for Conda {@link Repository} {@link Configuration}
+ */
+@CompileStatic
+trait CondaRepoRecipes
+    extends ConfigurationRecipes
+{
+  @Nonnull
+  Repository createCondaProxy(final String name,
+                            final String remoteUrl)
+  {
+    createRepository(createProxy(name, 'conda-proxy', remoteUrl))
+  }
+
+  abstract Repository createRepository(final Configuration configuration)
+}
