+++ nexus-repository-conda-it/src/test/java/org/sonatype/nexus/plugins/conda/internal/fixtures/RepositoryRuleConda.groovy
@@ -1,0 +1,15 @@
+package org.sonatype.nexus.plugins.conda.internal.fixtures
+
+import javax.inject.Provider
+
+import org.sonatype.nexus.repository.manager.RepositoryManager
+import org.sonatype.nexus.testsuite.testsupport.fixtures.RepositoryRule
+
+class RepositoryRuleConda
+  extends RepositoryRule
+  implements CondaRepoRecipes
+{
+  RepositoryRuleConda(final Provider<RepositoryManager> repositoryManagerProvider) {
+    super(repositoryManagerProvider)
+  }
+}
