--- src/main/java/org/sonatype/nexus/plugins/conda/internal/security/CondaFormatSecurityContributor.java
@@ -1,36 +1,0 @@
-/*
- * Sonatype Nexus (TM) Open Source Version
- * Copyright (c) 2019-present Sonatype, Inc.
- * All rights reserved. Includes the third-party code listed at http://links.sonatype.com/products/nexus/oss/attributions.
- *
- * This program and the accompanying materials are made available under the terms of the Eclipse Public License Version 1.0,
- * which accompanies this distribution and is available at http://www.eclipse.org/legal/epl-v10.html.
- *
- * Sonatype Nexus (TM) Professional Version is available from Sonatype, Inc. "Sonatype" and "Sonatype Nexus" are trademarks
- * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
- * Eclipse Foundation. All other trademarks are the property of their respective owners.
- */
-package org.sonatype.nexus.plugins.conda.internal.security;
-
-import javax.inject.Inject;
-import javax.inject.Named;
-import javax.inject.Singleton;
-
-import org.sonatype.nexus.plugins.conda.internal.CondaFormat;
-
-import org.sonatype.nexus.repository.Format;
-import org.sonatype.nexus.repository.security.RepositoryFormatSecurityContributor;
-
-/**
- * Conda format security resource.
- */
-@Named
-@Singleton
-public class CondaFormatSecurityContributor
-    extends RepositoryFormatSecurityContributor
-{
-  @Inject
-  public CondaFormatSecurityContributor(@Named(CondaFormat.NAME) final Format format) {
-    super(format);
-  }
-}
