--- src/main/resources/static/rapture/NX/conda/app/PluginStrings.js
@@ -1,36 +1,0 @@
-/*
- * Sonatype Nexus (TM) Open Source Version
- * Copyright (c) 2019-present Sonatype, Inc.
- * All rights reserved. Includes the third-party code listed at http://links.sonatype.com/products/nexus/oss/attributions.
- *
- * This program and the accompanying materials are made available under the terms of the Eclipse Public License Version 1.0,
- * which accompanies this distribution and is available at http://www.eclipse.org/legal/epl-v10.html.
- *
- * Sonatype Nexus (TM) Professional Version is available from Sonatype, Inc. "Sonatype" and "Sonatype Nexus" are trademarks
- * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
- * Eclipse Foundation. All other trademarks are the property of their respective owners.
- */
-
-/*global Ext, NX*/
-
-/**
- * Conda plugin strings.
- */
-Ext.define('NX.conda.app.PluginStrings', {
-  '@aggregate_priority': 90,
-
-  singleton: true,
-  requires: [
-    'NX.I18n'
-  ],
-
-  keys: {
-    Repository_Facet_CondaFacet_Title: 'Conda Settings',
-    SearchConda_Group: 'Conda Repositories',
-    SearchConda_License_FieldLabel: 'License',
-    SearchConda_Text: 'Conda',
-    SearchConda_Description: 'Search for components in Conda repositories',
-  }
-}, function(self) {
-  NX.I18n.register(self);
-});
